requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.3
markdown==3.5.1
requests
emoji==2.12.1
flask-cors
grpcio
grpcio-reflection
grpcio-health-checking
prometheus-flask-exporter
py-grpc-prometheus
opentelemetry-api
opentelemetry-sdk
opentelemetry-propagator-b3
opentelemetry-propagator-gcp
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
six
google-genai>=0.3.2
langchain>=0.1.0
langchain-google-vertexai>=2.0.11
google-cloud-compute>=1.19.0
google-cloud-resource-manager>=1.12.0